Добавил отображение активности follow пользователей: 1. like image 2. follow user 3. upload an image

+++ account/views.py
@@ -8,6 +8,8 @@ from django.views.decorators.http import require_POST
 from .forms import UserRegistrationForm, UserEditForm, ProfileEditForm, CustomAuthenticationForm
 from .models import Profile, Contact
 from common.decorators import ajax_required
+from actions.utils import create_action
+from actions.models import Action
 
 
 class CustomLoginView(LoginView):
@@ -16,7 +18,15 @@ class CustomLoginView(LoginView):
 
 @login_required
 def dashboard(request):
-    return render(request, 'account/dashboard.html', {'section': 'dashboard'})
+    # Display all actions by default
+    actions = Action.objects.exclude(user=request.user)
+    following_ids = request.user.following.values_list('id', flat=True)
+
+    if following_ids:
+        # If user is following others, retrieve only their actions
+        actions = actions.filter(user_id__in=following_ids)
+    actions = actions.select_related('user', 'user__profile').prefetch_related('target')[:10]
+    return render(request, 'account/dashboard.html', {'section': 'dashboard', 'actions': actions})
 
 
 @login_required
@@ -47,6 +57,7 @@ def register(request):
             )
             new_user.save()
             Profile.objects.create(user=new_user)
+            create_action(new_user, 'has created an account')
             return render(request, 'account/register_done.html', {'new_user': new_user})
 
     else:
@@ -73,11 +84,12 @@ def user_detail(request, username):
 def user_follow(request):
     user_id = request.POST.get('id')
     action = request.POST.get('action')
-    if user_id and action:
+    if user_id and action and user_id != str(request.user.id):
         try:
             user = User.objects.get(id=user_id)
             if action == 'follow':
                 Contact.objects.get_or_create(user_from=request.user, user_to=user)
+                create_action(request.user, 'is following', user)
             else:
                 Contact.objects.filter(user_from=request.user, user_to=user).delete()
             return JsonResponse({'status': 'ok'})
